Update MATH2319_Project_Phase1.Rmd
I've just added blank chunks and basic descriptions of tasks we could try. Also if you install.packages(TinyTeX) and run the pdf knit for this, it will produce a toc (table of contents, shown in the first chunk) which is dictated by the #, ##, ### for depth. It's hyperlinked in the pdf and looks profesh af.